用例管理平台/TestCasePla/requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11
django-crispy-forms==1.7.2
django-excel==0.0.10
django-formtools==2.1
django-import-export==1.1.0
django-reversion==3.0.2
DjangoUeditor==1.8.143
docopt==0.6.2
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.0
idna==2.8
jdcal==1.4
lml==0.0.9
odfpy==1.4.0
openpyxl==2.5.12
pipreqs==0.4.9
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-webio==0.1.4
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.12.0
tablib==0.12.1
texttable==1.5.0
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
yarg==0.1.9
